chore: moved permission func into utils
