+++ src/authorizations/utils/permissions.ts
@@ -135,3 +135,73 @@ export enum BucketTab {
   AllBuckets = 'All Buckets',
   Scoped = 'Scoped',
 }
+
+export const formatPermissionsObj = (permissions) => {
+
+  const newPerms = permissions.reduce((acc, {action, resource}) => {
+    const {type, id, orgID, name} = resource
+    let p
+
+    if (acc.hasOwnProperty(type)) {
+      p = {...acc[type]}
+      if (id && (type === 'buckets' || type === 'telegrafs')) {
+        if (p.sublevelPermissions.hasOwnProperty(id)) {
+          p.sublevelPermissions[id].permissions[action] = true
+        } else {
+          p.sublevelPermissions[id] = {
+            id: id,
+            orgID: orgID,
+            name: name,
+            permissions: {
+              read: action === 'read',
+              write: action === 'write',
+            },
+          }
+        }
+      } else {
+        p[action] = true
+      }
+    } else {
+      if (id && (type === 'buckets' || type === 'telegrafs')) {
+        p = {
+          read: false,
+          write: false,
+          sublevelPermissions: {
+            [id]: {
+              id: id,
+              orgID: orgID,
+              name: name,
+              permissions: {
+                read: action === 'read',
+                write: action === 'write',
+              },
+            },
+          },
+        }
+      } else {
+        p = {
+          read: action === 'read',
+          write: action === 'write',
+        }
+      }
+    }
+
+    return {...acc, [type]: p}
+  }, {})
+
+  Object.keys(newPerms).map(resource => {
+    const p = {...newPerms[resource]}
+    if (p.sublevelPermissions) {
+      p.read = !Object.keys(p.sublevelPermissions).some(
+        key => p.sublevelPermissions[key].permissions.read === false
+      )
+
+      p.write = !Object.keys(p.sublevelPermissions).some(
+        key => p.sublevelPermissions[key].permissions.write === false
+      )
+
+      newPerms[resource] = p
+    }
+  })
+  return newPerms
+}
